EmbeddingGenerator/requirements.txt
python-dotenv
pymongo
openai
